\name{BH184-package}
\alias{BH184-package}
\alias{BH184}
\docType{package}
\title{\packageTitle{BH184}}
\description{\packageDescription{BH184}}
\details{The DESCRIPTION file: \packageDESCRIPTION{BH184}}
\section{Package Content}{\packageIndices{BH184}}
\author{\packageAuthor{BH184}}
\section{Maintainer}{\packageMaintainer{BH184}}
\references{http://www.boost.org/}
\keyword{ package }
